backend/requirements.txt
-i https://pypi.org/simple
aiofiles==22.1.0
anyio==3.6.2 ; python_full_version >= '3.6.2'
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3 ; python_version >= '3.7'
cryptography==38.0.4
dnspython==2.2.1 ; python_version >= '3.6' and python_version < '4.0'
ecdsa==0.18.0 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==1.3.0
fastapi==0.88.0
greenlet==2.0.1 ; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
jinja2==3.1.2
markupsafe==2.1.1 ; python_version >= '3.7'
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic[email]==1.10.2
python-dotenv==0.21.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==1.4.45
starlette==0.22.0 ; python_version >= '3.7'
typing-extensions==4.4.0 ; python_version >= '3.7'
uvicorn==0.20.0
